zenz-mining-bot/app/bot.py
# app/bot.py  
import logging  
from telegram import Update  
from telegram.ext import Updater, CommandHandler, CallbackContext  
from mining_logic import get_mining_info  
import config  

logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s', level=logging.INFO)  
logger = logging.getLogger(__name__)  

def start(update: Update, context: CallbackContext) -> None:  
    update.message.reply_text('Welcome to the Web3 Mining Bot! Type /mining to check mining details.')  

def mining(update: Update, context: CallbackContext) -> None:  
    mining_info = get_mining_info()  
    update.message.reply_text(mining_info)  

def main():  
    updater = Updater(config.TELEGRAM_TOKEN)  
    dispatcher = updater.dispatcher  
    dispatcher.add_handler(CommandHandler('start', start))  
    dispatcher.add_handler(CommandHandler('mining', mining))  
    updater.start_polling()  
    updater.idle()  

if __name__ == '__main__':  
    main()

    # app/bot.py  
import logging  
from telegram import Update  
from telegram.ext import Updater, CommandHandler, CallbackContext  
from .config import TELEGRAM_TOKEN  

# Configure logging  
logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)  

def start(update: Update, context: CallbackContext):  
    update.message.reply_text("Welcome to the Web3 Mining Bot!")  

def mining(update: Update, context: CallbackContext):  
    # Logic for mining command (placeholder)  
    update.message.reply_text("Mining information...")   

def start_bot():  
    updater = Updater(token=TELEGRAM_TOKEN, use_context=True)  

    dp = updater.dispatcher  
    dp.add_handler(CommandHandler("start", start))  
    dp.add_handler(CommandHandler("mine", mining))  

    updater.start_polling()  
    updater.idle()